continue to remove dead code

+++ examples/perceptive_locomotion/systems/alip_mpfc_meshcat_visualizer.cc
@@ -71,7 +71,7 @@ void AlipMPFCMeshcatVisualizer::DrawFootholds(ConvexPolygonSet& foothold_set,
       drake::geometry::Rgba(0, 1, 0, 0.5),
       drake::geometry::Rgba(0, 0, 1, 0.5)
   };
-  for (int i = 0; i < foothold_set.size(); i++) {
+  for (size_t i = 0; i < foothold_set.size(); i++) {
     auto foothold = foothold_set.polygons().at(i);
     const auto [verts, faces] = foothold.GetSurfaceMesh();
     auto faces_reversed = faces;
